#25 Refactor pysafe_config structure to hide private imports
