+++ tests/test_helper_str_to_int.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pysafe_config._helper_int import _str_to_int
+from pysafe_config._internal._helpers import _str_to_int
 
 
 @pytest.mark.parametrize(
